# person_dashboard-main/Workbench/shared/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Personal AI Workbench Local API
  version: 1.5.0
  description: >-
    Local-loopback MVP contract. Foundation session/security and the durable
    project, milestone, task, manual discussion-to-decision, explicit text/link
    capture, daily focus/review, controlled Markdown source intake, and authorized
    context full-text search slices are implemented; other product domains remain out of scope.
servers:
  - url: http://127.0.0.1:8787
paths:
...
  /api/v1/session/bootstrap:
    post:
      operationId: bootstrapLocalSession
      parameters:
        - in: header
          name: X-Workbench-Bootstrap
          required: true
          schema: { type: string, minLength: 32 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: false }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "403": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /api/v1/session:
    get:
      operationId: getLocalSession
      security: [{ localSession: [] }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "401": { $ref: "#/components/responses/Error" }
    delete:
      operationId: revokeLocalSession
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: false }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/system/capabilities:
    get:
      operationId: getCapabilities
      security: [{ localSession: [] }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/spaces:
    get:
      operationId: listSpaces
      security: [{ localSession: [] }]
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "401": { $ref: "#/components/responses/Error" }
...
  /api/v1/sources:
    get:
      operationId: listSources
      description: Lists source metadata only after space/project authorization; storage paths and source bodies are not returned.
      security: [{ localSession: [] }]
      parameters:
        - in: query
          name: space_id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: project_id
          schema: { type: string, format: uuid }
        - in: query
          name: status
          schema: { type: string, enum: [ready, archived, failed], default: ready }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/sources/imports/markdown:
    post:
      operationId: importMarkdownSource
      description: >-
        Accepts one explicit Markdown payload from the browser, writes immutable content-addressed
        bytes under the controlled source root, creates Source/SourceVersion/Document, and indexes
        normalized text. It never accepts a client path, fetches a URL, invokes AI, or parses XLSX.
      security: [{ localSession: [] }]
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MarkdownSourceImport" }
      responses:
        "201": { $ref: "#/components/responses/Success" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/context/search:
    post:
      operationId: searchAuthorizedContext
      description: >-
        Read-only search command. Space/project/type/date authorization filters are applied in the
        database query before titles, snippets, counts, or locators are returned. The current baseline
        is FTS5 trigram with a bounded short-query fallback; it has no semantic recall or reranking.
      security: [{ localSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContextSearch" }
      responses:
        "200": { $ref: "#/components/responses/Success" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    localSession:
      type: apiKey
      in: cookie
      name: workbench_session
  parameters:
    CsrfToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string, minLength: 32 }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, minLength: 16, maxLength: 200, pattern: "^[A-Za-z0-9._~-]+$" }
...
  responses:
    Success:
      description: Successful envelope.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SuccessEnvelope" }
    Error:
      description: Safe error envelope.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
...
    NullableUuid:
      type: [string, "null"]
      format: uuid
...
    MarkdownSourceImport:
      type: object
      additionalProperties: false
      required: [space_id, filename, content]
      properties:
        space_id: { type: string, format: uuid }
        project_id: { $ref: "#/components/schemas/NullableUuid" }
        filename:
          type: string
          minLength: 1
          maxLength: 180
          pattern: "^[^/\\\\\\x00-\\x1F\\x7F]+\\.(md|markdown)$"
        title: { type: string, minLength: 1, maxLength: 200 }
        content: { type: string, minLength: 1, maxLength: 1000000 }
    ContextSearch:
      type: object
      additionalProperties: false
      required: [space_id, q]
      properties:
        space_id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        q: { type: string, minLength: 2, maxLength: 200 }
        types:
          type: array
          minItems: 1
          maxItems: 4
          uniqueItems: true
          default: [project, task, capture, document]
          items: { type: string, enum: [project, task, capture, document] }
        from: { type: string, pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$" }
        to: { type: string, pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$" }
        limit: { type: integer, minimum: 1, maximum: 50, default: 20 }
    SuccessEnvelope:
      type: object
      additionalProperties: false
      required: [request_id, status, data, errors, meta]
      properties:
        request_id: { type: string, format: uuid }
        status: { const: ok }
        data: {}
        errors: { type: array, maxItems: 0 }
        meta:
          type: object
          required: [api_version]
          properties:
            api_version: { const: v1 }
          additionalProperties: true
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [request_id, status, data, errors, meta]
      properties:
        request_id: { type: string, format: uuid }
        status: { const: error }
        data: { type: "null" }
        errors:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required: [code, message, field, retryable]
            properties:
              code: { type: string }
              message: { type: string }
              field: { type: [string, "null"] }
              retryable: { type: boolean }
        meta:
          type: object
          required: [api_version]
          properties:
            api_version: { const: v1 }
          additionalProperties: true